requirements.txt
torch
scipy>=0.19.0
numpy>=1.12.1
pandas>=0.19.2
pyyaml
statsmodels
tensorflow>=1.3.0
torch
tables
future
sklearn
matplotlib
gpytorch
math